+++ src/dsp_tools/commands/xml_validate/deserialise_project.py
@@ -1,7 +1,7 @@
 import json
 from typing import Any
 
-from dsp_tools.commands.xml_validate.models.project import ListDeserialised
+from dsp_tools.commands.xml_validate.models.project_deserialised import ListDeserialised
 
 
 def _get_project_ontology() -> Any:
